move types so its not loaded as command
